backend/pyproject.toml
[tool.poetry]
name = "fastapi-auth-example"
version = "0.1.0"
description = ""
authors = ["bjarnthoregilsson <bjarnthoregilsson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
SQLAlchemy = "^1.3.22"
python-jose = "^3.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
